pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["fooqoo"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.3"
python-dotenv = "^0.13.0"
tweepy = "^3.8.0"
fire = "^0.3.1"
google-cloud-datastore = "^1.12.0"
grpcio = "^1.28.1"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"
autopep8 = "^1.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
